+++ test/EasyPost/ErrorTest.php
@@ -58,6 +58,34 @@ Field errors:
     }
 
     /**
+     * Tests that we assign properties of an error correctly when returned via the alternative format.
+     * NOTE: Claims (among other things) uses the alternative errors format.
+     */
+    public function testErrorAlternativeFormat(): void
+    {
+        TestUtil::setupCassette('errors/errorsAlternativeFormat.yml');
+
+        try {
+            $claimData = Fixture::basicClaimData();
+            $claimData['tracking_code'] = '123';  // Intentionally pass a bad tracking code
+
+            self::$client->claim->create($claimData);
+        } catch (ApiException $error) {
+            $this->assertEquals(404, $error->getHttpStatus());
+            $this->assertEquals('NOT_FOUND', $error->code);
+            $this->assertEquals('The requested resource could not be found.', $error->getMessage());
+            $this->assertEquals('No eligible insurance found with provided tracking code.', $error->errors[0]);
+            $this->assertEquals('{"error":{"code":"NOT_FOUND","errors":["No eligible insurance found with provided tracking code."],"message":"The requested resource could not be found."}}', $error->getHttpBody()); // phpcs:ignore
+
+            // We check that the pretty printed output is the same here, leave the odd formatting as
+            // it is here and do not auto-format the next few lines.
+            $error->prettyPrint();
+            $this->expectOutputString('NOT_FOUND (404): The requested resource could not be found.
+');
+        }
+    }
+
+    /**
      * Test error deserialization with an array of error.message
      */
     public function testErrorMessageArray(): void